+++ workflow/tests/hpxml_translator_test.rb
@@ -485,7 +485,7 @@ class HPXMLTest < MiniTest::Test
         next if log_line.include? 'DSE is not currently supported when calculating utility bills.'
       end
       if !hpxml.header.power_outage_periods.empty?
-        next if log_line.include? 'Power outage period(s) may contain nonzero values'
+        next if log_line.include? 'It is not possible to eliminate all desired end uses (e.g. crankcase/defrost energy, water heater parasitics) in EnergyPlus during a power outage.'
       end
 
       flunk "Unexpected run.log warning found for #{File.basename(hpxml_path)}: #{log_line}"